feat: create a new file to avoid cirular deps
