+++ src/components/ItemEditorPage/CenterPanel/CenterPanel.tsx
@@ -4,7 +4,7 @@ import { Dropdown, DropdownProps, Popup, Icon } from 'decentraland-ui'
 import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import ViewPort from 'components/ViewPort'
 import { AvatarAnimation, PreviewType } from 'modules/editor/types'
-import { getSkinColors, getEyeColors, getHairColors } from 'modules/editor/utils'
+import { getSkinColors, getEyeColors, getHairColors } from 'modules/editor/colors'
 import { WearableBodyShape } from 'modules/item/types'
 import AvatarColorDropdown from './AvatarColorDropdown'
 import { Props, State } from './CenterPanel.types'
