+++ src/ecsScene/scene.ts
@@ -1,26 +1,14 @@
-// tslint:disable
 import { EventEmitter } from 'events'
-import {
-  engine,
-  GLTFShape,
-  Transform,
-  Entity,
-  Component,
-  NFTShape,
-  IEntity,
-  Vector3,
-  AvatarShape,
-  Color4,
-  Wearable
-} from 'decentraland-ecs'
+import { engine, GLTFShape, Transform, Entity, Component, NFTShape, IEntity, Vector3, AvatarShape, Wearable } from 'decentraland-ecs'
 import * as ECS from 'decentraland-ecs'
 import { createChannel } from 'decentraland-builder-scripts/channel'
 import { createInventory } from 'decentraland-builder-scripts/inventory'
-
 import { DecentralandInterface } from 'decentraland-ecs/dist/decentraland/Types'
 import { EntityDefinition, AnyComponent, ComponentData, ComponentType, Scene } from 'modules/scene/types'
 import { AssetParameterValues } from 'modules/asset/types'
 import { BODY_SHAPE_CATEGORY, WearableBodyShape } from 'modules/item/types'
+import { getEyeColors, getHairColors, getSkinColors } from 'modules/editor/colors'
+
 const { Gizmos, SmartItem } = require('decentraland-ecs') as any
 declare var dcl: DecentralandInterface
 
@@ -53,12 +41,12 @@ const scriptInstances = new Map<string, IScript<any>>()
 
 @Component('org.decentraland.staticEntity')
 // @ts-ignore
-export class StaticEntity { }
+export class StaticEntity {}
 
 @Component('org.decentraland.script')
 // @ts-ignore
 export class Script {
-  constructor(public assetId: string, public src: string, public values: AssetParameterValues) { }
+  constructor(public assetId: string, public src: string, public values: AssetParameterValues) {}
 }
 
 const editorComponents: Record<string, any> = {}
@@ -84,23 +72,23 @@ function getScriptInstance(assetId: string) {
   return instance
     ? Promise.resolve(instance)
     : scriptPromises
-      .get(assetId)!
-      .then(code => eval(code))
-      .then(() => load(assetId))
-      .then(Item => {
-        const instance = new Item()
-        scriptInstances.set(assetId, instance)
-        return instance
-      })
-      .catch(error => {
-        console.error(error.message)
-        // if something fails, return a dummy script
-        console.warn(`Failed to load script for asset id ${assetId}`)
-        return {
-          init() { },
-          spawn() { }
-        }
-      })
+        .get(assetId)!
+        .then(code => eval(code))
+        .then(() => load(assetId))
+        .then(Item => {
+          const instance = new Item()
+          scriptInstances.set(assetId, instance)
+          return instance
+        })
+        .catch(error => {
+          console.error(error.message)
+          // if something fails, return a dummy script
+          console.warn(`Failed to load script for asset id ${assetId}`)
+          return {
+            init() {},
+            spawn() {}
+          }
+        })
 }
 
 // avatar
@@ -116,9 +104,9 @@ function getAvatar(): Entity {
       'urn:decentraland:off-chain:base-avatars:',
       'dcl://base-avatars/'
     )
-    avatarShape.skinColor = new Color4(0.8671875, 0.6953125, 0.5625, 1)
-    avatarShape.hairColor = new Color4(0.8671875, 0.6953125, 0.5625, 1)
-    avatarShape.eyeColor = new Color4(0.8671875, 0.6953125, 0.5625, 1)
+    avatarShape.skinColor = getSkinColors()[0]
+    avatarShape.hairColor = getHairColors()[0]
+    avatarShape.eyeColor = getEyeColors()[0]
     avatarShape.name = 'Builder Avatar'
     avatarShape.wearables = []
     avatar.addComponent(avatarShape)
@@ -230,7 +218,6 @@ async function handleExternalAction(message: { type: string; payload: Record<str
       avatarShape.eyeColor = message.payload.eyeColor
       avatarShape.skinColor = message.payload.skinColor
 
-
       if (!avatar.isAddedToEngine()) {
         engine.addEntity(avatar)
       }
@@ -300,7 +287,7 @@ function createEntities(entities: Record<string, EntityDefinition>) {
 
     if (!entity) {
       entity = new Entity(builderEntity.name)
-        ; (entity as any).uuid = id
+      ;(entity as any).uuid = id
 
       if (!builderEntity.disableGizmos) {
         entity.addComponentOrReplace(gizmo)
